@@ -1,9 +1,8 @@
-import React from 'react';
-import { Grid, Button } from '@material-ui/core';
-import { Field, Formik, Form } from 'formik';
+import { Button, Grid } from '@mui/material';
+import { Field, Form, Formik } from 'formik';
 
-import { TextField, SelectField, GenderOption } from './FormField';
-import { Gender, Patient } from '../types';
+import { Gender, GenderOption, Patient } from '../types';
+import { SelectField, TextField } from '../components/FormField';
 
 /*
  * use type Patient, but omit id and entries,
