@@ -1,11 +1,13 @@
-import ReactDOM from 'react-dom';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 
 import App from './App';
 import { reducer, StateProvider } from './state';
 
-ReactDOM.render(
-  <StateProvider reducer={reducer}>
-    <App />
-  </StateProvider>,
-  document.getElementById('root')
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+  <React.StrictMode>
+    <StateProvider reducer={reducer}>
+      <App />
+    </StateProvider>
+  </React.StrictMode>
 );
