@@ -1,10 +1,56 @@
-export interface DiagnoseEntry {
+export interface Diagnosis {
   code: string;
   name: string;
   latin?: string;
 }
 
-export interface Entry {}
+interface BaseEntry {
+  id: string;
+  description: string;
+  date: string;
+  specialist: string;
+  diagnosisCodes?: Array<Diagnosis['code']>;
+}
+
+export interface BaseTypedEntry extends BaseEntry {
+  type: string;
+}
+
+export type NewBaseTypedEntry = Omit<BaseTypedEntry, 'id'>;
+
+export enum HealthCheckRating {
+  Healthy = 0,
+  LowRisk = 1,
+  HighRisk = 2,
+  CriticalRisk = 3,
+}
+
+export interface HealthCheckEntry extends BaseEntry {
+  type: 'HealthCheck';
+  healthCheckRating: HealthCheckRating;
+}
+
+export interface HospitalEntry extends BaseEntry {
+  type: 'Hospital';
+  discharge?: {
+    date: string;
+    criteria: string;
+  };
+}
+
+export interface OccupationalHealthcareEntry extends BaseEntry {
+  type: 'OccupationalHealthcare';
+  employerName: string;
+  sickLeave?: {
+    startDate: string;
+    endDate: string;
+  };
+}
+
+export type Entry =
+  | HealthCheckEntry
+  | HospitalEntry
+  | OccupationalHealthcareEntry;
 
 export enum Gender {
   Male = 'male',
@@ -20,6 +66,12 @@ export interface Patient {
   entries: Entry[];
 }
 
-export type PublicPatient = Omit<Patient, 'ssn' | 'entries'>;
+type UnionOmit<T, K extends string | number | symbol> = T extends unknown
+  ? Omit<T, K>
+  : never;
 
-export type NewPatient = Omit<Patient, 'id'>;
+export type PublicPatient = Omit<Patient, 'ssn'>;
+
+export type NewPatient = Omit<Patient, 'id' | 'entries'>;
+
+export type NewEntry = UnionOmit<Entry, 'id'>;
