@@ -4,6 +4,62 @@ export interface Diagnosis {
   latin?: string;
 }
 
+interface BaseEntry {
+  id: string;
+  description: string;
+  date: string;
+  specialist: string;
+  diagnosisCodes?: Array<Diagnosis['code']>;
+}
+
+export enum HealthCheckRating {
+  'Healthy' = 0,
+  'LowRisk' = 1,
+  'HighRisk' = 2,
+  'CriticalRisk' = 3,
+}
+
+interface HealthCheckEntry extends BaseEntry {
+  type: 'HealthCheck';
+  healthCheckRating: HealthCheckRating;
+}
+
+interface HospitalEntry extends BaseEntry {
+  type: 'Hospital';
+  discharge?: {
+    date: string;
+    criteria: string;
+  };
+}
+
+interface OccupationalHealthcareEntry extends BaseEntry {
+  type: 'OccupationalHealthcare';
+  employerName: string;
+  sickLeave?: {
+    startDate: string;
+    endDate: string;
+  };
+}
+
+export type Entry =
+  | HealthCheckEntry
+  | HospitalEntry
+  | OccupationalHealthcareEntry;
+
+export interface EntryFormValues extends Omit<BaseEntry, 'id'> {
+  type: string;
+  discharge?: {
+    date: string;
+    criteria: string;
+  };
+  employerName?: string;
+  sickLeave?: {
+    startDate: string;
+    endDate: string;
+  };
+  healthCheckRating?: HealthCheckRating;
+}
+
 export enum Gender {
   Male = 'male',
   Female = 'female',
@@ -17,6 +73,7 @@ export interface Patient {
   gender: Gender;
   ssn?: string;
   dateOfBirth?: string;
+  entries: Entry[];
 }
 
 export type Action =
@@ -27,8 +84,36 @@ export type Action =
   | {
       type: 'ADD_PATIENT';
       payload: Patient;
+    }
+  | {
+      type: 'UPDATE_PATIENT';
+      payload: Patient;
+    }
+  | {
+      type: 'SET_DIAGNOSIS_LIST';
+      payload: Diagnosis[];
+    }
+  | {
+      type: 'ADD_DIAGNOSIS';
+      payload: Diagnosis;
     };
 
 export type State = {
-  patients: { [id: string]: Patient | undefined };
+  patients: { [id: string]: Patient };
+  diagnoses: { [code: string]: Diagnosis };
+};
+
+export type GenderOption = {
+  value: Gender;
+  label: string;
+};
+
+export interface TypeOption {
+  value: string;
+  label: string;
+}
+
+export type HealthRatingOption = {
+  label: string;
+  value: number;
 };
